requirements.txt
# OneSpace Blog Backend Dependencies
# Python >= 3.10

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
aiomysql>=0.2.0
sqlalchemy[asyncio]>=2.0.0

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
bleach>=6.1.0

# Rate Limiting
slowapi>=0.1.9

# Utils
python-dotenv>=1.0.0
aiofiles>=23.2.0

# Logging
loguru>=0.7.0
